Add capability to use reserved or committed resources for SRVs and UAVs (#604)
This PR is the second step in bringing support to testing with partially
mapped and unmapped resources.
It lays out the infrastructure necessary to initialize UAV resources
with only a certain number of tiles mapped. By default, if the UAV is
specified as reserved, a sufficient number of tiles to cover the entire
size of the UAV resource will be mapped, unless otherwise specified.
The `IsReserved` keyword was added to the yaml resource setup, and if
not specified, the resources will be created as committed. The
TilesMapped keyword requires that the IsReserved keyword is set to True.

Addresses https://github.com/llvm/offload-test-suite/issues/182
Fixes https://github.com/llvm/offload-test-suite/issues/181
Fixes https://github.com/llvm/offload-test-suite/issues/623
The last part needed to complete the above issue (182) would be to
implement partial mappings for Buffers, but that will be left to a
separate PR.